requirements.txt
Flask==3.1.1
rasterio==1.3.9
numpy==1.26.4
Pillow==10.3.0
geopandas==0.14.3
gunicorn==23.0.0